user is now able to see the songs associated with an artist, and add songs. Updated styling and fixed a bug on artists link in nav bar
