+++ src/main/resources/templates/fragments/nav.html
@@ -17,7 +17,7 @@
                         <a class="nav-link" href="/">Home</a>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link" href="artists">Artists</a>
+                        <a class="nav-link" href="/artists">Artists</a>
                     </li>
                     <li class="nav-item">
                         <a class="nav-link" href="/logout">Logout</a>
